Update rules to handle newly numeric fields
Turns out we were already casting them in several places.

+++ dataactvalidator/config/sqlrules/c23_award_financial_4.sql
@@ -4,7 +4,7 @@ SELECT
 	(SELECT COALESCE(SUM(sub_af.transaction_obligated_amou::numeric),0) AS transaction_sum
 		FROM award_financial as sub_af
 		WHERE submission_id = {0} AND sub_af.uri = af.uri) AS transaction_obligated_amou_sum,
-  (SELECT COALESCE(SUM(sub_afa.federal_action_obligation::numeric),0) AS obligation_sum
+  (SELECT COALESCE(SUM(sub_afa.federal_action_obligation),0) AS obligation_sum
 		FROM award_financial_assistance as sub_afa
 		WHERE submission_id = {0} AND sub_afa.uri = af.uri and
 		COALESCE(sub_afa.assistance_type,'') not in ('07','08')) AS federal_action_obligation_sum,
@@ -21,7 +21,7 @@ GROUP BY af.uri
 HAVING
 		(SELECT COALESCE(SUM(sub_af.transaction_obligated_amou::numeric),0) AS transaction_sum
 		FROM award_financial as sub_af WHERE submission_id = {0} AND sub_af.uri = af.uri) <>
-		(-1*(SELECT COALESCE(SUM(sub_afa.federal_action_obligation::numeric),0) AS obligation_sum
+		(-1*(SELECT COALESCE(SUM(sub_afa.federal_action_obligation),0) AS obligation_sum
 		FROM award_financial_assistance as sub_afa
 		WHERE submission_id = {0} AND sub_afa.uri = af.uri and COALESCE(sub_afa.assistance_type,'') not in ('07','08')) -
 		(SELECT COALESCE(SUM(sub_afa.original_loan_subsidy_cost::numeric),0) AS obligation_sum
